Safari-609.1.11.1/LayoutTests/http/wpt/fetch/fetch-request-arraybuffer-content-type.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function(t) {
  let request = new Request("http://localhost", { method: "POST", body: new ArrayBuffer() });
  assert_equals(request.headers.get('Content-Type'), null);
}, "Content-Type header for payload of ArrayBuffer type should be null");
</script>
</body>
</html>
